# Optimise Apply[{ Either, \/ }].apply2