--- a/core/src/main/scala/scalaz/IList.scala
+++ b/core/src/main/scala/scalaz/IList.scala
@@ -434,6 +434,18 @@ sealed abstract class IList[A] extends Product with Serializable {
   def toZipper: Option[Zipper[A]] =
     sToZipper(toStream)
 
+  // optimised specialisation of Traverse.traverse for disjunction, avoiding
+  // allocations for each element, approximately 20% faster than the default
+  // implementation.
+  def traverseDisjunction[E, B](f: A => E \/ B): E \/ IList[B] = \/-(
+    map { a =>
+      f(a) match {
+        case -\/(err) => return -\/(err)
+        case \/-(j)   => j
+      }
+    }
+  )
+
   def uncons[B](n: => B, c: (A, IList[A]) => B): B =
     this match {
       case INil() => n