close #1956 

Running this perf test

```scala
jsonformat/jmh:run -i 5 -wi 5 -f1 -t1 -w1 -r1 .*Geo.*decodeManual*
```

from https://gitlab.com/fommil/scalaz-deriving

BEFORE this change

```
Benchmark                               Mode  Cnt      Score     Error  Units
GeoJSONBenchmarks.decodeManualError    thrpt    5  25727.854 ± 263.755  ops/s
GeoJSONBenchmarks.decodeManualSuccess  thrpt    5  26943.126 ± 230.573  ops/s
```

AFTER this change

```
GeoJSONBenchmarks.decodeManualError    thrpt    5  70845.254 ± 890.444  ops/s
GeoJSONBenchmarks.decodeManualSuccess  thrpt    5  84736.783 ± 473.980  ops/s
```

and WITH custom short circution impl of traverseDisjunction from #1956 (which is used in the repo's master... back out that change to run the above tests)

```
GeoJSONBenchmarks.decodeManualError    thrpt    5  586733.762 ±  6389.296  ops/s
GeoJSONBenchmarks.decodeManualSuccess  thrpt    5  110961.246 ±  468.384  ops/s
```

async-profiler alloc tracing shows that the `\/-` allocations are quite heavy. I guess that's to be expected.