@jdegoes 2018-08-13
@fommil Awesome work. Would be great to do this for `bind` too.