@fommil 2018-08-13
@jdegoes done, that actually has a non-negligible impact on my perf tests which aren't even testing for this!

```
[info] Benchmark                               Mode  Cnt      Score     Error  Units
[info] GeoJSONBenchmarks.decodeManualError    thrpt    5  74902.151 ± 488.520  ops/s
[info] GeoJSONBenchmarks.decodeManualSuccess  thrpt    5  88754.010 ± 759.063  ops/s
```