@fommil 2018-08-13
https://typelevel.org/blog/2014/11/10/why_is_adt_pattern_matching_allowed.html#adts-are-ok-to-go suggests that these optimisations are unneeded. My perf tests suggest otherwise on 2.12...