@edmundnoble 2018-08-13
You have quite a few "optimizations" here, not just touching `apply2`. Is it clear which are actually responsible for speedup, and why?