@fommil 2018-08-13
@edmundnoble the numbers cited are just the `\/.apply2`, everything else seems to have improved things a little bit and are just opportunistic while I'm in the area.